requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
botocore==1.22.10
cachetools==4.2.2
certifi==2021.10.8
jmespath==0.10.0
pynamodb==5.1.0
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
